requirements/base.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-jazzmin==2.6.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==24.1
phonenumbers==8.13.30
pillow==10.2.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
uritemplate==4.1.1
